♻️(backend) base views factorization
In order to make the `app_data` payload and view context more consistent
between views and also improve code path readability between views
with or without LTI request, this commit groups various common methods
inside common classes.

This is not perfect as the `BaseResourceView` is also used by the
`LTISelectView` which is quite far from the concerns of "specific"
resources, still it is supposed to make changes between views more explicit.