🚚(front) move document-specific components to relevant folder
Move components only used in the DashboardDocument to
components/DashboardDocuments/ instead of letting it in the root folder.Create
a new component DashboardDocumentPaneButtons which is a document-specialized
version of DashboardPaneButtons. This way it will be possible to get rid of
DashboardPaneButtons component, when it will not be used anymore by the
DashboardVideoPane (which will happen when DashboardVideo and its sub-
components will be removed).