✨(frontend) create a component to factorise all widget (live or vod)
Dashboards (live and vod) use similar layout with widget to configure the
resource under the main content. Those widgets used to be split and shared
between two components.
We rather create in single component to handle all widget and depend on a prop
to configure widgets for live or vod dashboard.