👷(ci) install Python 3.10 for e2e tests
We cannot update the playwright image to jammy to use Python 3.10 in end to
end tests, so we install Python 3.10 on the focal one.

Running webkit tests in `playwright:jammy` results in a timeout
see the same issue: https://github.com/microsoft/playwright/issues/15764

So we cannot update to `playwright:jammy` until the underlying OS
is `Ubuntu 20.04.4 LTS` (see the `Operating System` in the
"Spin up environment" step).