✨(front) manage metadata request with react-query
We want to fetch metadata from some resources, first for the video one.
We have to make an OPTIONS request and manage it using react-query. Once
the metadata fetch there is no reason at all to refetch them again
during the application lifecycle, also there is no reason to refresh it
at all so all these options are disable but can still be configured when
the useVideoMetadata hook is used in a component.