♻️(frontend) create libraries to share marsha components
In anticipation to create a standalone web app for marsha, we want to be able
to use components used in the lti application. In order to do that, we want to
create a module containing our components shared between both web app.
Taking advantage of yarn workspaces, the project organization is now around two
packages:
- app : containing web app to package and deploy
  - lti_site : lit web app
- packages : containing libs shared between web apps builded using rollup
  - lib_components : containing components shared between web apps
  - lib_tests : containing utilities functions to tests our components, this
had to be split in a dedicated lib to prevent those functions to be packaged by
rollup and imported in a webapp
  - lib_common : anything to be used in a web app, lib_components or lib_tests.
This imply that anything in the lib may be pure declarative statements and may
not be tested.