✨(frontend) authenticate the user in the SPA
Force the user to be authenticated to access website by redirecting him on the
authentication form on the backend.
When authenticated, the user is redirected back on the SPA with a token to
validate against the backend.
On challenge success, the backend return the user access token to fetch the api.