🏗️(frontend) project structure by features
Project structure by features is a way to organize your project by features.
It means that you will have a folder for each feature of your application.
Each feature folder will contain all the files related to this feature.
This structure is very useful when you have a lot of features
 in your application.
It is also very easy to find a file in this structure.