✨(backend) manage classroom document default attribute
As we want to send only one document as default presentation to BBB, the
first one created in a classroom it set to true, then if another one in
the same classroom is set to true, others are set to false.
This ensures that only one document per classroom is set as default.