🐛(back) do not use view.get_object in permissions
Since DRF 3.14.0 it is not possible to use view.get_object in the
has_permission method. If we do it, we have a max recursion error. In
this version, they fix a bug related to the usage of th OR and AND
operator. To fix this issue, we created a dedicated mixin responsible to
fetch the related object we are working on in the database.
https://github.com/encode/django-rest-framework/pull/7522