♻️(back) expose all resources we can work with lti select
On the lti-select endpoint, we want to expose all the resources
available an instructor can work with. The resource list is built
dynamically. All marsha apps are inspected and if a model lti_select is
found. In this module, a function called get_lti_select_config is
responsible to expose all the config needed to build the lti select
response.