♻️(front) dynamically build lti-select dashboard
The lti-select dashboard is now dynamically build based on what contains
appData added in the django view. If a targeted_resource is specified,
then only this resource is display and the user can only works on this
resource. Most of the existing component are extracted to their own
directory and reused depending if we are in Tabs context or a
targeted_resource context.