📝(frontend) add standalone site startup
This adds few notes on own to start the standalone frontend.

I also tried to update the makefile but did not manage to run the
standalone site inside a Node docker because it requires access to
the backend which would request more work (and I don't know if it's
wanted to add it to makefile).

The error is this one:
Proxy error: Could not proxy request /api/auth/challenge/
 from localhost:3000 to http://localhost:8060/.