✨(back) resource not ready and portability
When a resource is not ready, it should not raise a portability error for
intructors or administrators when playlist is ported.

This will prevent the future case where an instructor uploads many
videos from the standalone site and copy links in the LMS,
request portability, accept it while the video are still uploading.