🐛(frontend) role should not be tested in frontend
The Markdown editor was displayed if the LTI JWT provided the
`can_update` permission along with instructor or admin role,
but since the `can_update` attribute can only be true for
administrator and instructors it was not mandatory to test it.

Moreover it would have raise an issue for the standalone site.