✨(front) deploy site app in production
The site application will be serve by django in production on the same
domain name. For this, the build must be made in the django static and
the site view will serve the generate index file modified to replace
relative path to main js and css file by the django static url. A
postbuild script is created to copy all needed media static at the root
of django static directory.