🚨(front) install eslint from main package.json file
We are using the same version of eslint and its plugins everywhere in
marsha. Eslint and some plugins where present in every packages
node_moduls and it didn't find some plugins because it expects to find
them in the same place. Some where in the main node_modules and others
in the node_modules of each package. So solve this issue, we are using
the resolution in the main package and then use * everywhere eslint is
used.