✨(back) enrich playlist serializer with organization and consumer site
The playlist serializer had only organization and consumer site primary
key. We want to enrich these properties to contain more information
about these resources. An organization lite serializer is created for
this and a consumer site is also created. Moreoever, a dedicated
serializer for playlist creation is created to manage easily relations
with organization and consumer site.