✨(circle) set version to current commit sha1 when no tag
When we build the version.json file added in the docker image, the
version is set with the tag name. But if we have publishing the master
image, there is no tag version. To improve the usage of this master
image, we choose to substitute the tag name by the last commit sha1.
This will also improve debug, we will have the exact commit sha1
reported to sentry for example.