♻️(back) change playlist filter on list action to match permissions
In the permissions, a user can create, retrieve or update a playlist if
he is a plyalist administrator or if he his an orgnization administrator
of the playlist belonging to. On the list endpoint, only user member of
an organization can list these playlists, this does not reflect the
permissions. This endpoint is modified to return the playlists belonging
to an organization where the user is administrator but also all the
playlist where the user is administrator no matter the organization
relation.