🐛(frontend) do not lose search params during authentication
Authentication is based on a redirection on the backend auth page.
To supports link resource, the address the user is trying to reach is saved in
the local storage to redirect him onto when he is logged in.
But, we only stored location without search params, we now stores both.