♻️(front) rework frontend translations workflow
Translation are only for now in the lti_site apps. We have to rework all
the translation workflow to be able to translate standalone_site app and
all packages.
To avoid to translate multiple times the same translation or group them
on only one files, source translations are split in 3 files. One for the
lti_site, one for standalone_site and one for all packages containing
translations (This is all extract-* commands).
Then, to compile them, we use the compile-translations.js script to
merge the packages translations with the lti_site or the
standalone_site.