♻️(back) change startTime used to filter aws logs using live started_at
The management command check_live_state do its job by fetching aws logs.
To do limit the dataset filtered, we use the startTime parameter
available in the filter_log_events method. This startTime was the video
creation date but it's difficult to still use it because the video
creation date and the timestamp the live is effectively started can
differ a lot (several days/weeks). To be more accurate and fetch the
good events, we can restrict the startTime parameter to the effective
started_at timestamp. In this case events are always return.