✨(back) add management command clean_aws_elemental_stack
Once a live started, all AWS elemental stack are created. Once stopped, the
instructor must do an action. Restart it and/or convert it in VOD. If
nothing is done, the AWS element resources are leaved unused and use the quota
we have on our AWS account.
This resources unused must be removed after several days of inactivity and the
video move in a DELETED state.