🐛(back) fix permission for classroom creation
When a classroom is created, we don't know yet its primary key. So using
the permission IsClassroomPlaylistOrOrganizationAdmin works only for the
first classroom created, after creating a classroom fails. We have to
use the core_permission IsParamsPlaylistAdmin or
IsParamsPlaylistAdminThroughOrganization only on the create endpoint. To
avoid this mistake again, user_is and classroom_id paramter are now
checked.