🐛(back) fix permission for filedepository creation
When a filedepository is created, we don't know yet its primary key.
So using the permission IsFileDepositoryPlaylistOrOrganizationAdmin
works only for the first filedepository created, after creating a
new one fails. We have to use the core_permission IsParamsPlaylistAdmin
or IsParamsPlaylistAdminThroughOrganization only on the create endpoint.
To avoid this mistake again, user_id and classroom_id paramter are now
checked.