🐛(back) fix permission for markdown_document creation
When a markdown_document is created, we don't know yet its primary key.
So using the permission IsMarkdownDocumentPlaylistOrOrganizationAdmin
works only for the first markdown_document created. After, creating a
new one fails. We have to use the core_permission IsParamsPlaylistAdmin
or IsParamsPlaylistAdminThroughOrganization only on the create endpoint.
To avoid this mistake again, user_id and classroom_id paramter are now
checked.