✨(website) update classroom
Feature to update classroom from website.
The component to update the classroom comes from the package lib-classroom.
It is the same component that is used in LTI.
In order to work, we had to use a bit the same logic as in LTI
to initialise DashboardClassroom.
To do so, we had to wrap it in the same context,
and initialise it the same way.
The images being proper to LTI, we duplicated them in the website,
then to pass them in argument of our context.