♻️(back) set lti_id and consumer_site nullable on playlist model
We must be able to create a playlist without defining a lti_id or a
consumer_site. Both are only used in a LTI context and there is no
reason on the stansalone site to ask for this to an instructor when he
wants to create a new playlist. An integrity constrait is added to
ensure that lti_id ans consumer_site are still defined at the same time.