🐛(front) add classroom parameter on classroomdocument creation
In the POST request sent to create a classroomdocument, the `classroom`
paramter is not set. In a LTI context it is not needed, the classroom id
is retrieve in the LTI token. In the standalone site, this
information in not in the JWT token and must be set in the POST payload.
The backend is already dealing with this constraint, only the front
application must be updated.