🚨(front) adapt test to grommet 2.28.0
In grommet 2.28.0 they fix the issue #6159 allowing to choose a date
outside bounded range. We fix this case in our test by displaying an
error message when the date is in the past. We change the test to ensure
that is not possible to choose the date outside this range.