🐛(front) deal with webinar without thumbnail
In the VideoPlayer component, we are looking for thumbnail to use tham
as poster in the video element. For a webinar, there is no thumbnail nor
resolution available and sometimes the rendering of this component fails
on this part of the code. We are managing this case to avoid future
error.