➖(front) remove react-tooltip package
In the raw live dashboard, we were using the package react-tooltip to
display a success or error message when a user copy a RTMP link. This
package is only used in this component, in all other components we are
using a toast. We remove this package and use a toast instead.