🐛(front) guess ISO15897 locale code in standalone site
In the standalone site, the language code is search in the
navigator.language property. With this language code we try to load the
translation file dynamically. But the files have a specific name base on
a ISO15897 locale code and the value return navigator.language can be
not compatible with this format. To fix this issue, we have to maintain
a list of available react locales and thentry to guess this format from
the navigatore language.