✨(backend) add refresh access token endpoint
This add the endpoint to request new access token from
a refresh token.

The refresh token rotation is enabled, so a new one is
created anytime this endpoint is called. The previous
refresh token is blacklisted.

Note: this commit introduces new migration because of the
`token_blacklist` new application (and models).