✨(website) service worker token refresh
The service worker is now able to refresh the token when it is expired.
The worker listens to the fetch event and if the request is a 401, it
tries to refresh the token. If the refresh is successful, the request
is retried, if the refresh fails, the user is redirected to the login.