🐛(front) create unique storage name for useJwt hook
The zustand hook useJwt is using a persistent middleware. A unique name
must be set and this name is the key used as local storage key. In  LTI
context, we can have several iframe launched in the same page and the
JWT token will be overwritten between LTI app present in the page. To
fix this, we choose to create unique storage name by checking of in the
DOM we have the data-context set by the LTI view. If present we know we
are in a LTI context and we can pick in the LTI context the resource
modl and it's unique id.