🐛(frontend) set unique storage name for useJwt hook
The zustand hook useJwt is using a persistent middleware. A unique name
must be set and this name is the key used as local storage key. In  LTI
context, we can have several iframe launched in the same page and the
JWT token will be overwritten between LTI app present in the page. To
fix this, we set a unique storage name depend the resource id
of the application.