♻️(front) create resource_id in AppData
In the AppData, the resource property was removed once a copy made in
the model property. We need to know the resource id outside any react
context in order to configure the useJwt zustand storage name. This
storage name needs to be unique and we want to use the resource id to
make it unique. We create a new property resource_id in appData in order
to save the current resource id.