♻️(front) use appData.resource_id for useJwt store name uniqueness
The useJwt zustand store storage name must be unique. To do that, we use
the appData modelName and resource id. But in the ParseDataElement
function, the resource property is deleted. To fix that, we choose to
copy the resource id in a new property called resource_id. Then this new
property is used to create a unique name.