✨(back) create refresh token dedicated to marsha use cases
We want to manage several refresh tokens like we are managing several
access tokens. While it is possible to use multiple access token, the
refresh token endpoint is not able to do that. But the way a refresh
token build an access token from an existing refresh token (and it's the
case on the refresh token endpoint) allow us to keep all the simple-jwt
mechanism just by creating our own MarshaRefreshToken. This refresh
token will add a new property access_token_type in its payload and will
use it when it generates the access token. The MarshaRefreshToken must be
used in the TokenRefreshSerializer.