✨(frontend) display renater IDP's logo on login
Now the logo is provided in the IDP data, we can display
it on the login page.
Choice has been made to display an empty image when no logo
is provided to keep label aligned.

Size of the logo is enforced to 16x16 pixel as per the
Renater WAIF example.