🐛(frontend) fix label input too big for the input
Label in formfield is set in an absolute position to prevent the input in the
field (mostly select) from beeing ofset in the container. But, with long label,
the forfield may not resize accordingly and so the label would go beyond the
formfield.
Update the css to remove the position absolute but preventing the field from
expending in height the formfield and preserving the margins.