🐛(frontend) fix renater select issue on pagination
There is a bug in grommetjs select that could cause the site to crash.
This bug is due to pagination issue, with a useEffect in InfiniteScroll
component beeing called repetitively and calling each time a setState.
We are hiding this by setting the page size to the item list length.