🔥(frontend) remove service worker from standalone site and lti
The service worker was not installed everytime on the client side,
the refresh token could not be refreshed and the user was logged out.
So we decided to use a fetch wrapper instead of a service worker to
refresh the token.
This commit removes the service worker from the standalone site and
the lti site.
We remove it as well from the lib_components build.