🔧(tray) set limit to marsha pod resources
We first separate job and pod resources. We want to set limits on pod
but not on jobs. On pods, we set limits, configurable in each customer
environment. This limit is set to 1 for CPU because marsha needs lot of
CPU to start and the memory is set to 2.5Gb, it's a quite hight but this
is what we need for now.