♻️(tray) create a cronjob pipeline
The list of cronjob in marsha is growing and the template used is every
time the same. When we want to add a new cronjob, we copy/paste an
existing one and we modify the properties we want. Instead of doing
this, we can create a CronJobList resource and define the cronjobs in a
marsha_cronjobs variable define in each environment we want.