♻️(back) enable drf_spectacular url only in development
We want to expose drf_spectacular views only in development environment.
We decided to do that because there is a memory leak[1] in drf_spectacular
and this memory leak is impacting Marsha in production. The dockerflow
library expose an endpoint to check if the container is up and running.
To do this, it uses django check process and drf_spectacular generates a
full schema in its check. So the memory is growing everytime a check is
made.

[1]: https://github.com/tfranzel/drf-spectacular/issues/597