♻️(back) allow public resource to be embedded in a iframe
Public resource view is protected with the clickjacking protection
middleware. We want to allow consumer_site owner to embedded resources
related to their consumer_site in an iframe on their domain and
subdomain. To do that, if a consumer_site exists, we deactivate the
clickjacking middleware and we configure the Content-Security-Policy
frame-ancestors with the consumer_site domain and all subdomains. If
there is no consumer_site, like for public resources contained in a
playlist managed on the standalone_site, then we don't allow to embedded
the resource in an iframe at all.