🐛(back) manage header cache on public resources view
In commit b6b993b when a public resource is fetch and own in a
consumer_site, we changed the header to allow to embedded it in an
iframe using the CSP frame-ancestor header. But we didn't managed the
cache. So on the first request, we query the DB to retrieve the video
and then we check if there is a consumer_site. But then, for the next
requests we have a cache and we didn't query the DB anymore and we loose
the consumer_site information and the headers are unchanged. To fix
this, we save the domain in a dedicated cache and we use this cache to
determine if the current resource is owned in a consumer_site and
use its domain to change the response headers.