♻️(docker) change USER management in Dockerfile
We are having issue when running the development docker image for
linting the backend. The image is running with the root user, everytime
a file is modified we have to change this file ownership because it
belongs to root user. We don't want to use the root user anymore but an
Un-privileged one.