♻️(back) split xapi in LTI and website context
Depending from where the request is coming, a LTI context or from the
standalone site, we have to adapt how the xapi statement is computed. We
keep the log and xapi request logic but the statements are not enrich
the same way. From the standlone site context, we can use the user
information from the django database and the homepage for the actor
property comes from the django sites application.