💄(website) prioritize manual thumbnail over automated one
When a video is uploaded on our website, we automatically generate a
thumbnail for it. We can also manually select a thumbnail for a video.
If a video has both an automated and a manual thumbnail, we want to
prioritize the manual one and display it in the related video component.