♻️(back) do not share thumbnail instance in video serializer list
When a list of video is serialized, the first one having a thumbnail
stores ths thumbnail instance in a serializer property. This property is
never reset and the consequence is that this thumbnail instance will be
used by the next video instance serialized. To fix the issue, we must
force to set the thmubnail_instance to None at the beginning of the
to_representation method.