♻️(back) add refresh token in the challenge response
The response for the challenge request was not returning a refresh token
but only an access one. Every user connecting with shibboleth have no
refresh token and are deconnected fron the standalone site once the
access token is expired. So fix this, the challenge response has now a
refresh token.