🐛(website) fix disconnection when multiple tabs
When multiple tabs are open on the website,
if a tab stays inactive during 5mn (access token lifetime),
when it gets the focus back we are disconnected.
It is because the access token is not valid anymore,
and the refresh token has been updated by the other tab.
To fix this, we directly use localStorage to retrieve the jwt,
to be sure to always use an up to date JWT.