⚡️(back) control number of database queries on video API
The video model is the bigger one we have with lot of relations with
other model. The number of database queries made when a video or a
collection of video is fetch can grow quickly if we don't control them.
In this commit improvements are made to decrease the number of database
queries and to control this number in the time.