🏷️(frontend) change the jwt type to a more generic one
The jwt type was specific to the lti token. We had to change it
to a more generic one to be able to use it with different tokens,
with the web token in this case.
We had to get some permissions directly from the ressources instead of
the jwt, because the jwt didn't have everytime the informations depend
the token type.