✨(front) force xmpp disconnection when unmounting live dashboard
On the standalone site we must manage chat connection and disconnection
otherwise the converse client connects one to a xmpp room and never
disconnects. So when you navigate between several webinars you are still
seeing the same chat. First we must force chat disconnection when the
LiveUpdate component is unmounted, we force the user to logout and to
destroy the existing session. Then, the chat history must be reset to
not see messages between webinars when a user navigate.