✨(front) manage chat user useParticipantsStore on disconnect
On the standalone site, when a user leaves a webinar and disconnects
from the chat, the useParticipantsStore zustand store is not cleanup and
the current user is not removed in it. To fix this issue, we save the
converse user jid in the same store and remove the participant from this
store using this converse user jid.