🐛(front) correctly fetch transcript content in TranscriptReader
The react-query hook useTranscriptReaderRequest didn't use a unique
query key. That means that when the url was changing, this new url was
not use because the stale time was not over. First, we have to use a
unique query key by adding the timedtextrack id to it. Secondly the
usage of the `OnSuccess` on that react-query hook is not the good
solution. While no new query is made, the onSuccess is called with an
empty data. To fix this issue, all the logic made on this onSuccess is
moved in a `useEffect` hook in the TranscriptReader responsible to
update the cues. And last, there is no reason to fetch again and again
the file content, it will never change. So we can set an inifinte cache
and stale time.