⬆️(back) upgrade to django 4.1
In django 4.1, the full_clean method also check unicity. We disable this
to keep the same behavior as before.
On the Video model, the check constraint live_type_check is refactor and
does not use the RawSql anymore.
Some default order are modified to match expected tests.
usage of django.utils.timezone.utc is removed.