✨(live) sync video and medialive states command
Once a live started, video states becomes indirectly linked to
medialive channel state. Their is a risk that sometimes,
states are not sync for an unknown reason, for example, the api was down
when aws was calling our callback to update video states.
This can let us with media channel "IDLE",
while in our side, the video state is "RUNNING"
In this case, the video becomes soft locked, the possible actions
for a "RUNNING" video will fail because the
medialive channel state is "IDLE".
This command iterates on medialive channel and ensure that linked
videos are state synced to avoid this kind of soft lock.