🐛(front) fix wrong usage of reduce index in xapiVideo mergeSegments
There is a bug in the XapiVideoStatement mergeSegments method. Once a
first overlapsed segment merge, the reduce index does not match anymore
the accumulator lenght because we removed existing index in the previous
iteration. To fix that and keep having valide merged segments at the end
we must use the current accumulator length.