✨(lib-video) iframe code to copy to clipboard
When the video is_public is toggle on, after the public link, we added
a box with the iframe code beeing easily copy/paste.
The iframe contains all the permissions needed to display the video
correctly.
The permissions are different depending on the type of the video,
if it is a webinar or just a vod (microphone, camera...).