✨(website) prop withLogoLink on the HeaderLight
We have some cases where we want to have a link to the login page
on the HeaderLight component.
This commit adds a prop to the HeaderLight to optionally
have a link to the logo.
Because of the implementation in MainLayout, we created the
wrapper component HeaderLightLink that will be prefill the
prop withLogoLink.