🐛(frontend) add global test setup file to craco config
We have 2 useJwt versions, one for the standalone site and one for the
LTI. The tests were loading the LTI version because the
window.use_jwt_persistence var was not init to true.
We add a testSetup file to the craco config to init
window.use_jwt_persistence to true.