✨(back) add a middleware to deal with social auth exception
The middleware provided by django social auth to redirect the user to an
error page when an authentification failure exception is raised does not
match our need and we can't customize it like we want. We want to
redirect the user to an error page, display the error message and we
don't want to lose the exception. For this, the exception is captured by
sentry, then the user is redirected and the message is added in the
query string.