♻️(back) restrict access to video list endpoint
The list video endpoint was open to everybody belonging to an
organization or playlist. The problemt we have it that we are not able
to handle user without can_edit access on the front application. To
match the front requirement we have to restrict the list endpoint result
to organisation administrator or playlist admin/instructor.