✨(back) create a public and instructor token in classroom model
We don't want to generate anymore a JWT for the invite and instructor
token. Instead we decided to generate, when a classroom is created, a
public and an instructor token, based on the secrets python module. This
token are exposed in the serializer like the old token, only when the
`is_admin` context is set to True.