✨(back) endpoint checking classroom invite token
We need an endpoint checking a classroom invite token validity. The
endpoint is /api/classrooms/{classroom_id}/invite/?token={invite_token}
If the token is missing a 400 status code is return. If the token is
valid, a resource token is return with the good permissions set. If the
token is invalid, a 404 status code is return.