✅(frontend) utils advanceJestTimersByTime
React 18 have a new feature, the Automatic Batching.
https://react.dev/blog/2022/03/29/react-v18#new-feature-automatic-batching
It means that it will batch the state updates and the effects
together, and it will run them in the same tick.
With jest.advanceTimersByTime, we can advance the timers by time,
but it will not advance the effects by time, so we have to replay
them in a loop until we get the expected result.