✨(video) p2p videojs plugin
Create a VideoJS Plugin enabling the P2P for HLS.
To do so, we usee the `p2p-media-loader-core` to handle the p2p part,
and `videojs-hlsjs-plugin` to do the VideoJS integration.
The library needs the "process" variable and due to react-script 5+,
we have to use the `process`  library and to define it in webpack.