🐛(back) fix clean_aws_elemental_stack command
The clean_aws_elemental_stack is failing when medialive channel exists,
but its associated video does not.
It is also not catching if a medialive channel is not referenced
by a live, thus unused.
When a medialive channel is unused,
we should delete the stack of this medialive channel,
The same way we did with sync_medialive_video