🐛(frontend) context problem state in some cases
After beeing mounted, a Provider cannot have is initial value updated,
so we had a useEffect to update the context value.
Problem, in some cases, the initial state of the context was overloading the
updated state, depend if the initial state was considered as mutated
by the useEffect on render or not (if the initial state was a deep
object by example).
To fix the problem we removed the useEffect that updated the initial
state of the context, if we want to update the context value, we have
to use the context value setter.