✨(webtorrent) implement bittorrent-tracker server
We want to use our own webtorrent tracker for the p2p usage. We decided
to use the bittorent-tracker lib. We have to create a node project
dedicated to this usage. The node project is an express application,
using websocket connection. The connection to the webtorrent tracker is
allowed if a valide JWT token is used in the url. The connection is made
once.