🧐(frontend) report grommet components
add a new config file for react-scanner to report the grommet
components in marsha frontend.
Ultimately, we would like to reduce the number of grommet components
that we use in marsha. This first step gives us a state of the art of
the current presence of it in our code.