🎨(backend) use cache for recordings url
the bbb api has changed and does not provide a permanent link for
recordings anymore.
The link provided has a 1h lifetime before being destroyed.
In this commit, we put the recording url in a cache while it is still
available and reftech it when too much time has gone.