✨(back) manage shibboleth metadata config per site
When the xml metadata are generated, we want to customize them depending
the site used. A new backend extending saml_fer is created, overriding
the generate_saml_config method. In this method, the request domain is
tested and if a site config exists, we will use it to customize the
return configuration.