🐛(back) remove stamp uses in aws stacks names
Using stamps in aws name stacks remove the possibility to use
"unique" names for resources. This is a problem when it twp requests
are made on the same video, but due to different stamps, two stacks
will be created.
Stamps are used elsewhere in the code, so we cannot remove them.
Therefore, we move this information to the 'Tags' properties.