♿️(frontend) remove redundant form role in the LTI app
It's considered as a bad A11y practice to add redundant semantic.
A `<form>` element has already a `form` role, specifying it again is
redundant. It might make screen readers behave buggy.

I would suggest adding the eslint-plugin-jsx-a11y plugin to the
ESLint custom configuration. It's a dependency of `react-scripts`,
used by the standalone app but not the LTI app.

Note: A <form> needs an accessible name to have the `form` role.