🐛(backend) xapi cache to avoid duplicate statements
Sometimes video xAPI statements are sent multiple times, it can
happens when a client has network connections issues and the
request is retried.
To avoid duplicate statements, we use the `id` field of the
statement to identify if the statement has already been sent
using a cache of a few minutes.