🔧(back) add new setting PLAYLIST_CLAIM_EXCLUDED_LTI_USER_ID
(back) add new setting  new setting PLAYLIST_CLAIM_EXCLUDED_LTI_USER_ID
is added to manage a list of lti_user_id to exclude. In edxapp studio,
the lti_user_id used is an anonymous one and has all the time the value
`student`. Creating an association with this value must be forbidden