🔨(docker) create a script managing docker compose binary to use
In the makefile, we made a modification allowing to use docker compose
in version 1 or 2. But we also use the `docker-compose` binary in
script present in the bin directory. To use the right binary, we created
a docker-compose script responsible to check the compose version and we
use this script everywhere we want to use docker compose.