🐛(back) filter classroomDocument queryset on the current classroom
In the ClassroomDocumentViewset, the queryset is not filtered on the
current classroom. The list action is now managed on this viewsete and
not on the ClassroomViewset since the playlist token refacto is made.
This lead to list all the documents existing in the database. To fix
this issue, all request are now filter using the current classroom.