🐛(back) filter depositedFile queryset on the current FileDepository
In the DepositedFileViewSet, the queryset is not filtered on the
current filedepository. The list action is now managed on this viewset
and not on the FileDepositoryViewset since the playlist token refacto
is made.
This lead to list all the deposited files existing in the database.
To fix this issue, all request are now filter using the current
file depository.