♻️(xapi) change downloading statement category related activity
Download a video or a document are statements that were not officially defined
in xAPI profiles. A new LMS profile release in the next weeks by Gaia-X DASES
organisation defines the spec for this statement.
The `category` field in context.contextActivities traces this information: it is
added for `download a document` statement definition. `Download a video`
statement was mistakenly related to the official Video profile IRI but was not
part of its specification. It is replaced with the LMS profile IRI.