🐛(front) display claim link only when is-claimed request succeeded
The claim link was display by default by setting the state based on
appDatata content. This can lead to displaying the link few second and
the state is changed only if the resource is-claimed. This introduce a
bug because this endpoint can not be fetch by a student, so student have
the link display. This fix does the opposite, when mounted, the state is
set to false and then the link is display only if the request succeed
and is claimed is strictly false.