♻️(lib-video) refacto LanguageSelect timed text track
We had some states flipping with the LanguageSelect component
with "no language available" option. The no language available option
was selectable, and would create an error in the upload process
if selected.
We remove the "no language available" option, and let the Cunningham
component handle the case when no options are available.