✨(frontend) create Typo component
Typo is a generic component that can be used to render any HTML tag,
it includes a set of predefined methods and styles adapted with the
Cunningham design system. It will make it easier to create new
specific html components and to maintain the consistency of the
design system.
We refacto Headings component to use Typo component.