🐛(back) force bbb user_id uniqueness
When a user is joining a classroom, created in the standalone site, from
an invite link, the JWT token
does not contains user id and the resource does not belong to a consumer
site. In that case, the user_id parameters set in the join url will be
the same for all user using an invite link. With BBB 2.6 it seems that
it is not possible anymore to have multiple time the same user with the
same user_id. Surprisingly user are ejected once 4 users with the same
user_id have joined and they are ejected randomly.